Implemented feedback on PR
