+++ src/odemis/util/__init__.py
@@ -19,77 +19,6 @@ PURPOSE. See the GNU General Public License for more details.
 You should have received a copy of the GNU General Public License along with
 Odemis. If not, see http://www.gnu.org/licenses/.
 
-
-The function `check_random_state()` is copied from scikit-learn
-utils/validation.py, which is licenced under the following terms and
-conditions:
-
-    BSD 3-Clause License
-
-    Copyright (c) 2007-2021 The scikit-learn developers.
-    All rights reserved.
-
-    Redistribution and use in source and binary forms, with or without
-    modification, are permitted provided that the following conditions are met:
-
-    * Redistributions of source code must retain the above copyright notice,
-      this list of conditions and the following disclaimer.
-
-    * Redistributions in binary form must reproduce the above copyright notice,
-      this list of conditions and the following disclaimer in the documentation
-      and/or other materials provided with the distribution.
-
-    * Neither the name of the copyright holder nor the names of its
-      contributors may be used to endorse or promote products derived from
-      this software without specific prior written permission.
-
-    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-    AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-    IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-    ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-    LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-    CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-    SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-    INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-    CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-    ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-    POSSIBILITY OF SUCH DAMAGE.
-
-
-The function `rng_integers()` is copied from scipy _lib/_util, which is
-licensed under the following terms and conditions:
-
-    Copyright (c) 2001-2002 Enthought, Inc.  2003-2019, SciPy Developers.
-    All rights reserved.
-
-    Redistribution and use in source and binary forms, with or without
-    modification, are permitted provided that the following conditions
-    are met:
-
-    1. Redistributions of source code must retain the above copyright
-       notice, this list of conditions and the following disclaimer.
-
-    2. Redistributions in binary form must reproduce the above
-       copyright notice, this list of conditions and the following
-       disclaimer in the documentation and/or other materials provided
-       with the distribution.
-
-    3. Neither the name of the copyright holder nor the names of its
-       contributors may be used to endorse or promote products derived
-       from this software without specific prior written permission.
-
-    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-    LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-    A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-    OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-    LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-    DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-    THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
 """
 
 # Various helper functions that have a generic usefulness
@@ -102,7 +31,6 @@ import inspect
 import itertools
 import logging
 import math
-import numbers
 import queue
 import signal
 import sys
@@ -119,118 +47,18 @@ from decorator import decorator
 
 from . import weak
 
-try:
-    from numpy.random import Generator as Generator
-except ImportError:
-
-    class Generator:  # type: ignore[no-redef]
-        pass
-
-
+# Used in the type signature of `pairwise()` below such that we can define the
+# return type as a function of the input type.
 T = TypeVar("T")
 
 
 # helper functions
 def pairwise(iterable: Iterable[T]) -> Iterable[Tuple[T, T]]:
-    "s -> (s0,s1), (s1,s2), (s2, s3), ..."
+    """s -> (s0,s1), (s1,s2), (s2, s3), ..."""
     # will be added to itertools in Python 3.10
     a, b = itertools.tee(iterable)
     next(b, None)
     return zip(a, b)
-
-
-def check_random_state(seed):
-    """
-    Turn `seed` into a `numpy.random.RandomState` instance.
-
-    Parameters
-    ----------
-    seed : {None, int, `numpy.random.Generator`, `numpy.random.RandomState`}, optional
-        If `seed` is None (or `numpy.random`), the `numpy.random.RandomState`
-        singleton is used.
-        If `seed` is an int, a new ``RandomState`` instance is used,
-        seeded with `seed`.
-        If `seed` is already a ``Generator`` or ``RandomState`` instance then
-        that instance is used.
-
-    Returns
-    -------
-    seed : {`numpy.random.Generator`, `numpy.random.RandomState`}
-        Random number generator.
-
-    """
-    if seed is None or seed is numpy.random:
-        return numpy.random.mtrand._rand
-    if isinstance(seed, (numbers.Integral, numpy.integer)):
-        return numpy.random.RandomState(seed)
-    if isinstance(seed, (numpy.random.RandomState, Generator)):
-        return seed
-
-    raise ValueError(
-        "%r cannot be used to seed a numpy.random.RandomState instance" % seed
-    )
-
-
-def rng_integers(gen, low, high=None, size=None, dtype="int64", endpoint=False):
-    """
-    Return random integers from low (inclusive) to high (exclusive), or if
-    endpoint=True, low (inclusive) to high (inclusive). Replaces
-    `RandomState.randint` (with endpoint=False) and
-    `RandomState.random_integers` (with endpoint=True).
-
-    Return random integers from the "discrete uniform" distribution of the
-    specified dtype. If high is None (the default), then results are from
-    0 to low.
-
-    Parameters
-    ----------
-    gen : {None, numpy.random.RandomState, numpy.random.Generator}
-        Random number generator. If None, then the numpy.random.RandomState
-        singleton is used.
-    low : int or array-like of ints
-        Lowest (signed) integers to be drawn from the distribution (unless
-        high=None, in which case this parameter is 0 and this value is used
-        for high).
-    high : int or array-like of ints
-        If provided, one above the largest (signed) integer to be drawn from
-        the distribution (see above for behavior if high=None). If array-like,
-        must contain integer values.
-    size : array-like of ints, optional
-        Output shape. If the given shape is, e.g., (m, n, k), then m * n * k
-        samples are drawn. Default is None, in which case a single value is
-        returned.
-    dtype : {str, dtype}, optional
-        Desired dtype of the result. All dtypes are determined by their name,
-        i.e., 'int64', 'int', etc, so byteorder is not available and a specific
-        precision may have different C types depending on the platform.
-        The default value is numpy.int_.
-    endpoint : bool, optional
-        If True, sample from the interval [low, high] instead of the default
-        [low, high) Defaults to False.
-
-    Returns
-    -------
-    out: int or ndarray of ints
-        size-shaped array of random integers from the appropriate distribution,
-        or a single such random int if size not provided.
-
-    """
-    if isinstance(gen, Generator):
-        return gen.integers(low, high=high, size=size, dtype=dtype, endpoint=endpoint)
-    else:
-        if gen is None:
-            # default is RandomState singleton used by np.random.
-            gen = numpy.random.mtrand._rand
-        if endpoint:
-            # inclusive of endpoint
-            # remember that low and high can be arrays, so don't modify in place
-            if high is None:
-                return gen.randint(low + 1, size=size, dtype=dtype)
-            if high is not None:
-                return gen.randint(low, high=high + 1, size=size, dtype=dtype)
-
-        # exclusive
-        return gen.randint(low, high=high, size=size, dtype=dtype)
 
 
 def get_best_dtype_for_acc(idtype, count):
